Fix a bunch of compilation warnings in VC++ due to size_t to int conversions.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6336 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ idlib/Dict.cpp
@@ -454,7 +454,7 @@ static void WriteString( const char *s, idFile *f ) {
 		idLib::common->Error( "idDict::WriteToFileHandle: bad string" );
 	}
 
-	f->Write( s, strlen(s) + 1 );
+    f->Write(s, static_cast<int>(strlen(s) + 1));
 }
 
 /*
@@ -536,7 +536,7 @@ const idKeyValue *idDict::MatchPrefix( const char *prefix, const idKeyValue *las
 
 	assert( prefix );
 
-	const int len = strlen( prefix );
+    const int len = static_cast<int>(strlen(prefix));
 	int start = -1;
 
 	if ( lastMatch ) {
